# static/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "component API",
    "version": "1.0.0",
    "description": "API for managing Bills of Materials (components)"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api"
    }
  ],
  "paths": {
...
    "/api/state/{component_id}": {
      "post": {
        "tags": [
          "state"
        ],
        "summary": "states of a component",
        "description": "Can perform action of changing a component state",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "owner and note of the state",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "owner": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User not logged in"
          }
        }
      },
      "get": {
        "tags": [
          "state"
        ],
        "summary": "state of a component",
        "description": "Can perform action of seeing a component state",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "owner and note of the state",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "owner": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  },
                  "component_id": {
                    "type": "integer"
                  },
                  "state": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User not logged in"
          }
        }
      }
    },
    "/state": {
      "get": {
        "tags": [
          "state"
        ],
        "summary": "states of all components",
        "description": "Can perform action of seeing all components states",
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "owner and note of the state",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "owner": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  },
                  "component_id": {
                    "type": "integer"
                  },
                  "state": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - User not logged in"
          }
        }
      }
    },
...
  }
}